♻️ (GEKExample): Update exercises with new validate changes
